RAG_Morocco_Project/requirements.txt
langchain-openai>=0.1.8
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
openai>=1.30.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pymupdf>=1.24.0
Pillow>=10.3.0
streamlit>=1.35.0
